long int=>long long, %ld=>%lld

+++ D7_pn-check.cpp
@@ -4,11 +4,11 @@
 #include <math.h> //to use math functions, e.g., sqrt, sqrtf, pow, powf
 #include <time.h> //for time
 
-long int check_prime (long int);
+long long check_prime(long long);
 
 int main(void) {
-    long int ix, isqrtx, i, ipf=0; //long int=18.9桁: %ld
-    time_t tstart, tend;    
+    long long ix, isqrtx, i, ipf = 0; //long int=18.9桁: %ld
+    time_t tstart, tend;
 
     tstart = time(NULL);
 
@@ -17,22 +17,22 @@ int main(void) {
     ipf = check_prime(ix);
 
     if(ipf==0){ //割り切れる数がなかった場合
-        printf("%ld is a prime number.\n", ix);
+        printf("%lld is a prime number.\n", ix);
     }else{
-        printf("%ld can be devided by %ld.\n", ix,ipf);
+        printf("%lld can be devided by %ld.\n", ix,ipf);
     }
 
     tend = time(NULL);
-    printf("time for computing=%ld [sec]\n", tend-tstart);
+    printf("time for computing=%lld [sec]\n", tend-tstart);
 
     return 0; 
 }
 
-long int check_prime(long int inum){
-    long int i, isqrt_num, ipfactor;
+long long check_prime(__int64 inum) {
+    long long i, isqrt_num, ipfactor;
 
     ipfactor=0;
-    isqrt_num=(long int)sqrt((double)inum);
+    isqrt_num = (long long)sqrt((double)inum);
 
     for(i=3; i<=isqrt_num; i=i+2){
         if((inum%i)==0){
